mirgrate to sound null safety

+++ experimental/federated_plugin/federated_plugin_platform_interface/lib/battery_method_channel.dart
@@ -12,6 +12,6 @@ class BatteryMethodChannel extends FederatedPluginInterface {
 
   @override
   Future<int> getBatteryLevel() async {
-    return await _methodChannel.invokeMethod<int>('getBatteryLevel');
+    return await _methodChannel.invokeMethod<int>('getBatteryLevel') as int;
   }
 }
